requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blinker==1.9.0
certifi==2025.7.14
click==8.2.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==2.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.24.4
openai==1.3.6
openpyxl==3.1.5
pandas==2.0.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
